Move GuestProfiler to the wasmtime crate
