+++ crates/wasmtime/Cargo.toml
@@ -35,6 +35,7 @@ cfg-if = "1.0"
 log = { workspace = true }
 wat = { workspace = true, optional = true }
 serde = { version = "1.0.94", features = ["derive"] }
+serde_json = { workspace = true }
 bincode = "1.2.1"
 indexmap = "1.6"
 paste = "1.0.3"
@@ -45,6 +46,7 @@ object = { workspace = true }
 async-trait = { workspace = true, optional = true }
 encoding_rs = { version = "0.8.31", optional = true }
 bumpalo = "3.11.0"
+fxprof-processed-profile = "0.6.0"
 
 [target.'cfg(target_os = "windows")'.dependencies.windows-sys]
 workspace = true
