+++ crates/wasmtime/src/lib.rs
@@ -399,6 +399,7 @@ mod limits;
 mod linker;
 mod memory;
 mod module;
+mod profiling;
 mod r#ref;
 mod signatures;
 mod store;
@@ -416,6 +417,7 @@ pub use crate::limits::*;
 pub use crate::linker::*;
 pub use crate::memory::*;
 pub use crate::module::Module;
+pub use crate::profiling::GuestProfiler;
 pub use crate::r#ref::ExternRef;
 #[cfg(feature = "async")]
 pub use crate::store::CallHookHandler;
